requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Base de données
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Cache & tâches async
redis==5.1.1
celery==5.4.0
celery[redis]==5.4.0

# Pipeline PDF
pdfplumber==0.11.4
requests==2.32.3
beautifulsoup4==4.12.3

# IA Claude (fallback parsing + offres)
anthropic==0.40.0

# Email
resend==2.4.0

# PDF génération (rapports)
reportlab==4.2.5

# NLP (parsing optionnel)
spacy==3.8.2

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilitaires
python-dotenv==1.0.1
httpx==0.27.2
